Added Proper Controller
